backend/requirements-dev.txt
fastapi
uvicorn
httpx[http2]
neo4j
pydantic-settings
redis
asyncpg
psycopg[binary]
qdrant-client
aiokafka
prometheus-client
python-dotenv
PyJWT>=2.9.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Security Testing Tools
bandit>=1.7.5
safety>=3.0.0
semgrep>=1.45.0
detect-secrets>=1.4.0
